requirements.txt
altgraph==0.17.5
anyio==4.12.0
arrow==1.4.0
beautifulsoup4==4.14.3
binaryornot==0.4.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.0
colorama==0.4.6
cookiecutter==2.6.0
cryptography==46.0.3
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pefile==2024.8.26
pillow==12.0.0
pycparser==2.23
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pypdfium2==5.1.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==7.4.2
repath==0.9.0
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==4.0.2
